requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
python-dotenv
stripe
dash
dash-core-components
dash-html-components
plotly
